Split out typeDefs and resolvers
